feat(history): rrevise history comments UI
