+++ src/views/Me/History/Comments/index.tsx
@@ -1,11 +1,172 @@
+import _flatten from 'lodash/flatten'
+import { useContext } from 'react'
 import { useIntl } from 'react-intl'
 
-import { Head, Layout } from '~/components'
+import {
+  analytics,
+  collapseContent,
+  filterComments,
+  mergeConnections,
+  toPath,
+} from '~/common/utils'
+import { Head, Layout, LinkWrapper, UserDigest } from '~/components'
+import {
+  ArticleDigestTitle,
+  EmptyComment,
+  InfiniteScroll,
+  List,
+  QueryError,
+  Spinner,
+  usePublicQuery,
+  ViewerContext,
+} from '~/components'
+import { MeCommentsQuery } from '~/gql/graphql'
 
 import HistoryTabs from '../HistoryTabs'
-import BaseUserComments from './UserComments'
+import { ME_COMMENTS } from './gql'
+import styles from './styles.module.css'
 
-const UserComments = () => {
+type CommentedArticleComment = NonNullable<
+  NonNullable<
+    NonNullable<
+      MeCommentsQuery['node'] & { __typename: 'User' }
+    >['commentedArticles']['edges']
+  >[0]['node']['comments']['edges']
+>[0]['node']
+
+type CommentArticle = NonNullable<
+  NonNullable<
+    NonNullable<
+      MeCommentsQuery['node'] & { __typename: 'User' }
+    >['commentedArticles']['edges']
+  >[0]['node']['comments']['edges']
+>[0]['node'] & { __typename: 'Article' }
+
+const Comments = () => {
+  const viewer = useContext(ViewerContext)
+
+  /**
+   * Data Fetching
+   */
+  // public data
+  const { data, loading, error, fetchMore } = usePublicQuery<MeCommentsQuery>(
+    ME_COMMENTS,
+    {
+      variables: { id: viewer?.id },
+    }
+  )
+
+  // pagination
+  const connectionPath = 'node.commentedArticles'
+  const { edges, pageInfo } =
+    (data?.node?.__typename === 'User' &&
+      data.node.commentedArticles &&
+      data.node.commentedArticles) ||
+    {}
+
+  // load next page
+  const loadMore = async () => {
+    analytics.trackEvent('load_more', {
+      type: 'user_comment',
+      location: edges?.length || 0,
+    })
+
+    await fetchMore({
+      variables: { after: pageInfo?.endCursor },
+      updateQuery: (previousResult, { fetchMoreResult }) =>
+        mergeConnections({
+          oldData: previousResult,
+          newData: fetchMoreResult,
+          path: connectionPath,
+        }),
+    })
+  }
+
+  /**
+   * Render
+   */
+  if (loading) {
+    return <Spinner />
+  }
+
+  if (error) {
+    return <QueryError error={error} />
+  }
+
+  if (!edges || edges.length <= 0 || !pageInfo) {
+    return <EmptyComment />
+  }
+
+  const articleEdges = edges
+    .map((edge) => {
+      const commentEdges = edge.node.comments.edges || []
+      const comments = filterComments<CommentedArticleComment>(
+        commentEdges.map(({ node }) => node)
+      )
+      return { ...edge, comments }
+    })
+    .filter(({ comments }) => comments.length > 0)
+
+  return (
+    <InfiniteScroll hasNextPage={pageInfo.hasNextPage} loadMore={loadMore}>
+      <List responsiveWrapper>
+        {articleEdges.map(({ cursor, node, comments }) => (
+          <List.Item key={cursor}>
+            <section className={styles.item}>
+              <section className={styles.user}>
+                <UserDigest.Mini
+                  user={node.author}
+                  avatarSize="sm"
+                  textSize="xs"
+                  hasAvatar
+                  hasDisplayName
+                />
+              </section>
+
+              <section className={styles.title}>
+                <LinkWrapper
+                  {...toPath({
+                    page: 'articleDetail',
+                    article: node,
+                  })}
+                >
+                  <ArticleDigestTitle
+                    article={node}
+                    is="h2"
+                    textSize="md"
+                    lineClamp={1}
+                  />
+                </LinkWrapper>
+              </section>
+
+              <List hasBorder={false}>
+                {comments.map((comment) => (
+                  <List.Item key={comment.id}>
+                    <section className={styles.comment}>
+                      <LinkWrapper
+                        {...toPath({
+                          page: 'commentDetail',
+                          comment,
+                          article: node as CommentArticle,
+                        })}
+                      >
+                        <section className={styles.content}>
+                          {collapseContent(comment.content)}
+                        </section>
+                      </LinkWrapper>
+                    </section>
+                  </List.Item>
+                ))}
+              </List>
+            </section>
+          </List.Item>
+        ))}
+      </List>
+    </InfiniteScroll>
+  )
+}
+
+const HistoryComments = () => {
   const intl = useIntl()
   const title = intl.formatMessage({
     defaultMessage: 'History',
@@ -22,9 +183,9 @@ const UserComments = () => {
 
       <HistoryTabs />
 
-      <BaseUserComments />
+      <Comments />
     </Layout.Main>
   )
 }
 
-export default UserComments
+export default HistoryComments
